pyproject.toml
[project]
name = "as-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "black>=25.1.0",
    "celery>=5.5.3",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pyright==1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-dotenv>=0.5.2",
    "redis>=6.2.0",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.41",
]
[tool.ruff]
line-length=100
